reworked entire code, finaly works

+++ StatusCommands.py
@@ -6,78 +6,59 @@ class StatusCommands(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command(name="available", aliases=["av"])
+    #standart status and RPC
+    @commands.command(name="ON", aliases=["on"])
     async def command_available(self, ctx):
-        """command_available gets called when the user executes the command '>available' or '>av' into any channel"""
         if self.bot.debug:
             print("command_available called.")
 
         await self.bot.change_presence(
-            status=discord.Status.do_not_disturb,
+            status=discord.Status.idle,
             activity=discord.Activity(
-                type=discord.ActivityType.watching,
-                name="cyka"
-            ),
-            afk=True
-        )
-
-        await ctx.message.delete()
-
-    @commands.command(name="sleep")
-    async def command_sleep(self, ctx):
-        """command_sleep gets called when the user executes the command '>sleep' into any channel"""
-        if self.bot.debug:
-            print("command_sleep called.")
-
-        await self.bot.change_presence(
-            status=discord.Status.do_not_disturb,
-            activity=discord.Activity(
-                type=discord.ActivityType.watching,
-                name="Cyka"
+                type=discord.ActivityType.playing,
+                name="with the shotgun trigger"
             ),
             afk=False
         )
 
-        await ctx.message.delete()
+        if ctx.message:
+            await ctx.message.delete()
 
-    @commands.command(name="school", aliases=["work"])
+    #sets status to invisible plus custom RPC
+    @commands.command(name="invisible", aliases=["off"])
+    async def command_sleep(self, ctx):
+        if self.bot.debug:
+            print("command_sleep called.")
+
+        await self.bot.change_presence(
+            status=discord.Status.invisible,
+            activity=discord.Activity(
+                type=discord.ActivityType.watching,
+                name="marks mom"
+            ),
+            afk=False
+        )
+
+        if ctx.message:
+            await ctx.message.delete()
+
+    #sets status to Do not Discturb plus custom RPC
+    @commands.command(name="Work", aliases=["worki"])
     async def command_school(self, ctx):
-        """command_school gets called when the user executes the command '>school' or '>work' into any channel"""
         if self.bot.debug:
             print("command_school called.")
 
         await self.bot.change_presence(
             status=discord.Status.do_not_disturb,
             activity=discord.Activity(
-                type=discord.ActivityType.watching,
-                name="Cyka"
+                type=discord.ActivityType.playing,
+                name="with my sanity"
             ),
-            afk=True
+            afk=False
         )
-
-        await ctx.message.delete()
-
-    @commands.command(name="homework", aliases=["hw"])
-    async def command_homework(self, ctx, *, is_afk: str = "no"):
-        """command_homework gets called when the user executes the command '>homework' or '>hw' into any channel"""
-        if self.bot.debug:
-            print("command_homework called.")
-
-        if is_afk == "yes":
-            is_afk = True
-        else:
-            is_afk = False
-
-        await self.bot.change_presence(
-            status=discord.Status.do_not_disturb,
-            activity=discord.Activity(
-                type=discord.ActivityType.watching,
-                name="cyka"
-            ),
-            afk=is_afk
-        )
-
-        await ctx.message.delete()
+        
+        if ctx.message:
+            await ctx.message.delete()
 
 
 async def setup(bot):
